rename: 3D 이미지 관련 컴퍼넌트 3D 폴더에 통합 #50
* /components/ -> /components/3D/
