+++ frontend-dashboard/src/components/3D/Model3D.jsx
