+++ frontend-dashboard/src/components/3D/ThreeDimensionalImage.jsx
